exceptions handlers

+++ src/main/java/com/paulinabury/demo/service/StudentServiceDbImpl.java
@@ -1,5 +1,6 @@
 package com.paulinabury.demo.service;
 
+import com.paulinabury.demo.exceptions.*;
 import com.paulinabury.demo.model.Student;
 import com.paulinabury.demo.repository.StudentRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -24,19 +25,40 @@ public class StudentServiceDbImpl implements StudentService {
 
 
     @Override
-    public Optional<Student> getStudentById(Long id) {
-        return studentRepository.findById(id);
+    public Optional<Student> getStudentById(Long id) throws WrongIdException {
+        if (studentRepository.existsById(id)) {
+            return studentRepository.findById(id);
+        } else {
+            throw new WrongIdException("Podano błędne id");
+        }
     }
 
     @Override
     public List<Student> getAllStudents() {
+        if (studentRepository.findAll().size() == 0)
+            throw new WrongIdException("Brak studentów do wyświetlenia");
         return studentRepository.findAll();
     }
 
     @Override
-    public boolean deleteStudentById(Long id) {
-        studentRepository.deleteById(id);
-        return true;
+    public boolean deleteStudentById(Long id) throws WrongIdException {
+        if (studentRepository.existsById(id)) {
+            studentRepository.deleteById(id);
+            return true;
+        } else {
+            throw new WrongIdException("Podano błędne id");
+        }
+    }
+
+    @Override
+    public boolean deleteStudentByIndexNumber(String index) throws WrongIndexNumberException {
+        Student student = selectStudentByIndexNumber(index);
+        if (student != null) {
+            studentRepository.deleteById(student.getId());
+            return true;
+        } else {
+            throw new WrongIndexNumberException("Podano błędny numer indeksu");
+        }
     }
 
     @Override
@@ -50,26 +72,40 @@ public class StudentServiceDbImpl implements StudentService {
     }
 
     @Override
-    public void updateStudentById(Long id, Student student) {
+    public void updateStudentById(Long id, Student student) throws WrongIdException {
         if (studentRepository.existsById(id)) {
             student.setId(id);
             studentRepository.save(student);
+        } else {
+            throw new WrongIdException("Podano błędne id");
         }
-
     }
 
     @Override
-    public List<Student> selectStudentByIndexNumber(String indexNumber) {
+    public Student selectStudentByIndexNumber(String indexNumber) throws WrongIndexNumberException {
+        if (studentRepository.selectAllStudentsWithIndexEqualTo(indexNumber) == null)
+            throw new WrongIndexNumberException("Podano błędny nr indeksu");
         return studentRepository.selectAllStudentsWithIndexEqualTo(indexNumber);
     }
 
     @Override
-    public List<Student> selectStudentByName(String name) {
+    public List<Student> selectStudentByName(String name) throws WrongNameException {
+        if (studentRepository.selectAllStudentsWithNameEqualsTo(name).size() == 0)
+            throw new WrongNameException("Brak studentów o podanym imieniu");
         return studentRepository.selectAllStudentsWithNameEqualsTo(name);
     }
 
     @Override
-    public List<Student> selectStudentByField(String field) {
+    public List<Student> selectStudentBySurname(String surname) throws WrongSurnameException {
+        if (studentRepository.selectAllStudentsWithSurnameEqualsTo(surname).size() == 0)
+            throw new WrongSurnameException("Brak studentów o podanym nazwisku");
+        return studentRepository.selectAllStudentsWithSurnameEqualsTo(surname);
+    }
+
+    @Override
+    public List<Student> selectStudentByField(String field) throws WrongFieldException {
+        if (studentRepository.selectAllStudentsWithFieldEqualsTo(field).size() == 0)
+            throw new WrongFieldException("Brak studentów na podanym kierunku");
         return studentRepository.selectAllStudentsWithFieldEqualsTo(field);
     }
